Main.java modified

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        long current = 2_000_000_000;
-        long transfer = 500_000_000;
-        long result = transfer + current;
+        int current = 2_000_000_000;
+        int transfer = 500_000_000;
+        int result = transfer + current;
 
         System.out.println(result);
